requirements.txt
streamlit
streamlit-extras
streamlit-antd-components
streamlit-lottie
requests
fastapi
uvicorn
python-multipart
pandas
openpyxl
xlrd
pytesseract
pdf2image
Pillow
opencv-python
numpy
spacy
bcrypt
faker
reportlab